Adding Application discriptor and schema yaml files
The application.yaml file which is been added under template folder is a application descriptor.This manifest describes the application and is used in the UI. \n The schema yaml file is used when we build your deployer and usage points are : \n  Information about the current release, such as its release version, release notes, etc. A declaration and parameterization of all images used by the app.\n The parameters that users can customize when they deploy the app, and how they are rendered in the deployment configuration UI.\n Cluster constraints and requirements, such as the minimum required version of Kubernetes.